Made `feedItems.hooks.ts`
